# api-backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IntelliQ API",
        "version": "0.0.1"
    },
    "paths": {
...
        "/intelliq_api/getquestionanswers/{questionnaireID}/{questionID}": {
            "get": {
                "summary": "Get answers for a given question",
                "description": "Returns the answers to a given question",
                "parameters": [
                    {
                        "in": "path",
                        "name": "questionnaireID",
                        "description": "The ID of the questionnaire",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "questionID",
                        "description": "The ID of the question",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "format",
                        "description": "The format to return the data in, either JSON or CSV",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "csv",
                                "json"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The answers to the question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "questionnaireID": {
                                            "type": "string"
                                        },
                                        "questionID": {
                                            "type": "string"
                                        },
                                        "answers": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "session": {
                                                        "type": "string"
                                                    },
                                                    "ans": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            },
                            "text/csv": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request. Returns if no sessionID or questionnaireID is provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "No data. Returns if no data is found for the given questionnaireID and questionID.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/intelliq_api/question/{questionnaireID}/{questionID}": {
            "get": {
                "summary": "Get data for a specific question",
                "description": "Returns data for a specific question.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "questionnaireID",
                        "required": true,
                        "description": "ID of the questionnaire the question belongs to"
                    },
                    {
                        "in": "path",
                        "name": "questionID",
                        "required": true,
                        "description": "ID of the question"
                    },
                    {
                        "in": "query",
                        "name": "format",
                        "description": "Format for the returned data (either \"csv\" or \"json\")"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully returned data for the specified question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "questionnaireID": {
                                            "type": "string",
                                            "description": "ID of the questionnaire the question belongs to"
                                        },
                                        "qID": {
                                            "type": "string",
                                            "description": "ID of the question"
                                        },
                                        "qtext": {
                                            "type": "string",
                                            "description": "Text of the question"
                                        },
                                        "required": {
                                            "type": "string",
                                            "description": "Indicates whether the question is required (either \"true\" or \"false\")"
                                        },
                                        "type": {
                                            "type": "string",
                                            "description": "Type of the question (either \"question\" or \"profile\")"
                                        },
                                        "options": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "optID": {
                                                        "type": "string",
                                                        "description": "ID of the option"
                                                    },
                                                    "opttxt": {
                                                        "type": "string",
                                                        "description": "Text of the option"
                                                    },
                                                    "nextqID": {
                                                        "type": "string",
                                                        "description": "ID of the next question"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No questionnaireID or questionID was provided"
                    },
                    "402": {
                        "description": "No data was found for the specified questionnaireID and questionID"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}